requirements.txt
asgiref==3.2.10
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cleo==0.7.6
clikit==0.4.3
contoml==0.32
cryptography==2.9.2
defusedxml==0.7.0rc1
distlib==0.3.1
Django==3.0
django-ckeditor==5.9.0
django-email-verification==0.0.6
django-js-asset==1.2.2
django-phone-verify==2.0.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
first==2.0.2
html5lib==1.1
idna==2.10
importlib-metadata==1.1.3
install-requires==0.3.0
iso8601==0.1.12
jeepney==0.4.3
Jinja2==2.11.2
jsonschema==3.2.0
keyring==20.0.1
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==1.0.0
mysqlclient==2.0.1
nexmo==2.4.0
oauthlib==3.1.0
packaging==20.4
pastel==0.2.0
pexpect==4.8.0
phonenumbers==8.12.6
Pillow==7.2.0
pkginfo==1.5.0.1
poetry==1.0.10
poetry-setup==0.3.6
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylev==1.3.0
pyparsing==2.4.7
pyrsistent==0.14.11
python-decouple==3.3
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
requirements-parser==0.2.0
requirementslib==1.0.11
SecretStorage==3.1.2
shellingham==1.3.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
strict-rfc3339==0.7
timestamp==0.0.1
toml==0.10.1
tomlkit==0.5.11
twilio==6.44.0
urllib3==1.25.9
webencodings==0.5.1
yapf==0.30.0
zipp==3.1.0
